Balance factor

+++ 14-binary_tree_balance.c
@@ -1,9 +1,14 @@
 #include "binary_trees.h"
 
+/**
+ * binary_tree_balance - measures the balance factor of a binary tree
+ * @tree: pointer to the root of the binary tree
+ * Return: balance factor of a b-tree
+ */
 int binary_tree_balance(const binary_tree_t *tree)
 {
-	size_t l = 0;
-	size_t r = 0;
+	int l = 0;
+	int r = 0;
 
 	if (!tree)
 		return (0);
